refactor: refactored Exceptions in Lists
