+++ datastax/Lists/DoublyCircularList.py
@@ -1,11 +1,10 @@
 from typing import Any
-
 from datastax.Lists.DoublyLinkedList import DoublyLinkedList
 from datastax.Lists.CircularLinkedList import CircularLinkedList
 from datastax.Lists.AbstractLists import DoublyCircularList as AbstractList
 
 
-class DoublyCircularList(AbstractList, CircularLinkedList, DoublyLinkedList):
+class DoublyCircularList(CircularLinkedList, DoublyLinkedList, AbstractList):
 
     def append(self, data: Any) -> None:
         super().append(data)
